refactor: fota
