+++ src/FilePicker/FilePicker.tsx
@@ -3,14 +3,12 @@ import React, { createRef } from 'react'
 export const FilePicker = ({
 	accept,
 	maxSize,
-	onError,
 	onFile,
 	...restProps
 }: {
 	[key: string]: any
 	accept: string
 	maxSize: number
-	onError: (result: Error) => void
 	onFile: (file: File) => void
 }) => {
 	const inputRef = createRef<HTMLInputElement>()
