+++ src/Settings/Settings.tsx
@@ -6,7 +6,7 @@ import { NumberConfigSetting } from './NumberConfigSetting'
 import { formatDistanceToNow } from 'date-fns'
 import { emojify } from '../Emojify/Emojify'
 import styled from 'styled-components'
-import { mobileBreakpoint } from '../Styles'
+import { mobileBreakpoint } from '../theme/bootstrap4/Styles'
 import {
 	DataModules,
 	DeviceConfig,
