+++ src/aws/Cats/Page.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { List } from './List'
-import { Main } from '../../Styles'
+import { Main } from '../../theme/bootstrap4/Styles'
 import { ErrorCard } from '../../theme/bootstrap4/Error'
 import { Progress } from '../../theme/bootstrap4/Progress'
 import { CatsList } from '../../theme/bootstrap4/CatsList'
