+++ src/aws/CatsMap/Page.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { CatsMap } from './Loader'
-import { CatMapMain } from '../../CatsMap/Container'
+import { CatMapMain } from '../../theme/bootstrap4/CatsMap/Container'
 
 export const CatsMapPage = () => (
 	<CatMapMain>
