+++ src/azure/About/Page.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { Configuration } from '../../theme/bootstrap4/azure/About/Configuration'
-import { Main } from '../../Styles'
+import { Main } from '../../theme/bootstrap4/Styles'
 import { User } from '../../theme/bootstrap4/azure/About/User'
 import { SolutionConfigConsumer, AccessTokenConsumer } from '../App'
 import { About } from '../../theme/bootstrap4/About/About'
