+++ src/azure/Cats/Page.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import { Main } from '../../Styles'
+import { Main } from '../../theme/bootstrap4/Styles'
 import { ApiClientConsumer } from '../App'
 import { List } from './List'
 import { ButtonWarnings } from '../../ButtonWarnings/ButtonWarnings'
